move data_directory fixture to top level conftest
