+++ conftest.py
@@ -1,6 +1,10 @@
 import fnmatch
 import os
 import sys
+import pytest
+
+from ibis.compat import Path
+
 
 collect_ignore = ['setup.py']
 
@@ -11,3 +15,17 @@ if sys.version_info.major == 2:
         for filename in filenames:
             if fnmatch.fnmatch(filename, '*.py'):
                 collect_ignore.append(os.path.join(root, filename))
+
+
+@pytest.fixture(scope='session')
+def data_directory():
+    root = Path(__file__).absolute().parent
+
+    default = root / 'ci' / 'ibis-testing-data'
+    datadir = os.environ.get('IBIS_TEST_DATA_DIRECTORY', default)
+    datadir = Path(datadir)
+
+    if not datadir.exists():
+        pytest.skip('test data directory not found')
+
+    return datadir
